etl/extract/smart_thermo_service/requirements.txt
boto3==1.26.41
pandas==1.5.2
kafka-python